+++ buildSrc/src/main/kotlin/software/ralf/app/platform/gradle/buildsrc/Plugins.kt
@@ -5,7 +5,6 @@ internal object Plugins {
   const val ANDROID_KMP_LIBRARY = "com.android.kotlin.multiplatform.library"
   const val ANDROID_LINT = "com.android.lint"
   const val APP_PLATFORM = "software.ralf.app.platform"
-  const val BINARY_COMPAT_VALIDATOR = "org.jetbrains.kotlinx.binary-compatibility-validator"
   const val COMPOSE_COMPILER = "org.jetbrains.kotlin.plugin.compose"
   const val COMPOSE_MULTIPLATFORM = "org.jetbrains.compose"
   const val DETEKT = "dev.detekt"